fix: renamed to extensions
