+++ advanced_alchemy/extensions/litestar/cli.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, cast
 from litestar.cli._utils import LitestarGroup, console
 from rich.prompt import Confirm, Prompt
 
-from advanced_alchemy.integrations.litestar.plugin import AlembicCommands, _get_advanced_alchemy_plugin
+from advanced_alchemy.extensions.litestar.plugin import AlembicCommands, _get_advanced_alchemy_plugin
 
 if TYPE_CHECKING:
     from litestar import Litestar
